major module reorg
