+++ docs/Language-PureScript-Interpreter-FFI-Internal.html
@@ -1,0 +1,1 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8" /><meta name="viewport" content="width=device-width, initial-scale=1" /><title>Language.PureScript.Interpreter.FFI.Internal</title><link href="linuwial.css" rel="stylesheet" type="text/css" title="Linuwial" /><link rel="stylesheet" type="text/css" href="quick-jump.css" /><link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=PT+Sans:400,400i,700" /><script src="haddock-bundle.min.js" async="async" type="text/javascript"></script><script type="text/x-mathjax-config">MathJax.Hub.Config({ tex2jax: { processClass: "mathjax", ignoreClass: ".*" } });</script><script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script></head><body><div id="package-header"><span class="caption">runpurs-hs-0.1.0.0</span><ul class="links" id="page-menu"><li><a href="src/Language.PureScript.Interpreter.FFI.Internal.html">Source</a></li><li><a href="index.html">Contents</a></li><li><a href="doc-index.html">Index</a></li></ul></div><div id="content"><div id="module-header"><table class="info"><tr><th>Safe Haskell</th><td>None</td></tr><tr><th>Language</th><td>Haskell2010</td></tr></table><p class="caption">Language.PureScript.Interpreter.FFI.Internal</p></div><div id="interface"><h1>Documentation</h1><div class="top"><p class="src"><a id="v:forAll" class="def">forAll</a> :: (SourceType -&gt; SourceType) -&gt; SourceType <a href="src/Language.PureScript.Interpreter.FFI.Internal.html#forAll" class="link">Source</a> <a href="#v:forAll" class="selflink">#</a></p></div><div class="top"><p class="src"><a id="v:array" class="def">array</a> :: SourceType -&gt; SourceType <a href="src/Language.PureScript.Interpreter.FFI.Internal.html#array" class="link">Source</a> <a href="#v:array" class="selflink">#</a></p></div><div class="top"><p class="src"><a id="v:function" class="def">function</a> :: SourceType -&gt; SourceType -&gt; SourceType <a href="src/Language.PureScript.Interpreter.FFI.Internal.html#function" class="link">Source</a> <a href="#v:function" class="selflink">#</a></p></div></div></div><div id="footer"><p>Produced by <a href="http://www.haskell.org/haddock/">Haddock</a> version 2.24.0</p></div></body></html>
