+++ docs/Language-PureScript-Interpreter-FFI.html
@@ -1,0 +1,1 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8" /><meta name="viewport" content="width=device-width, initial-scale=1" /><title>Language.PureScript.Interpreter.FFI</title><link href="linuwial.css" rel="stylesheet" type="text/css" title="Linuwial" /><link rel="stylesheet" type="text/css" href="quick-jump.css" /><link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=PT+Sans:400,400i,700" /><script src="haddock-bundle.min.js" async="async" type="text/javascript"></script><script type="text/x-mathjax-config">MathJax.Hub.Config({ tex2jax: { processClass: "mathjax", ignoreClass: ".*" } });</script><script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script></head><body><div id="package-header"><span class="caption">runpurs-hs-0.1.0.0</span><ul class="links" id="page-menu"><li><a href="src/Language.PureScript.Interpreter.FFI.html">Source</a></li><li><a href="index.html">Contents</a></li><li><a href="doc-index.html">Index</a></li></ul></div><div id="content"><div id="module-header"><table class="info"><tr><th>Safe Haskell</th><td>None</td></tr><tr><th>Language</th><td>Haskell2010</td></tr></table><p class="caption">Language.PureScript.Interpreter.FFI</p></div><div id="table-of-contents"><div id="contents-list"><p class="caption" onclick="window.scrollTo(0,0)">Contents</p><ul><li><a href="#g:1">Foreign function interface</a></li></ul></div></div><div id="synopsis"><details id="syn"><summary>Synopsis</summary><ul class="details-toggle" data-details-id="syn"><li class="src short"><span class="keyword">data</span> <a href="#t:FFI">FFI</a> m = <a href="#v:FFI">FFI</a> {<ul class="subs"><li><a href="#v:ffi_moduleName">ffi_moduleName</a> :: ModuleName</li><li><a href="#v:ffi_values">ffi_values</a> :: [<a href="Language-PureScript-Interpreter-FFI.html#t:ForeignImport" title="Language.PureScript.Interpreter.FFI">ForeignImport</a> m]</li></ul>}</li><li class="src short"><span class="keyword">data</span> <a href="#t:ForeignImport">ForeignImport</a> m = <a href="#v:ForeignImport">ForeignImport</a> {<ul class="subs"><li><a href="#v:fv_name">fv_name</a> :: Ident</li><li><a href="#v:fv_type">fv_type</a> :: SourceType</li><li><a href="#v:fv_value">fv_value</a> :: <a href="Language-PureScript-Interpreter-Types.html#t:Value" title="Language.PureScript.Interpreter.Types">Value</a> m</li></ul>}</li><li class="src short"><a href="#v:toEnv">toEnv</a> :: <a href="Language-PureScript-Interpreter-FFI.html#t:FFI" title="Language.PureScript.Interpreter.FFI">FFI</a> m -&gt; <a href="Language-PureScript-Interpreter-Types.html#t:Env" title="Language.PureScript.Interpreter.Types">Env</a> m</li><li class="src short"><a href="#v:toExterns">toExterns</a> :: <a href="Language-PureScript-Interpreter-FFI.html#t:FFI" title="Language.PureScript.Interpreter.FFI">FFI</a> m -&gt; ExternsFile</li></ul></details></div><div id="interface"><a href="#g:1" id="g:1"><h1>Foreign function interface</h1></a><div class="top"><p class="src"><span class="keyword">data</span> <a id="t:FFI" class="def">FFI</a> m <a href="src/Language.PureScript.Interpreter.FFI.html#FFI" class="link">Source</a> <a href="#t:FFI" class="selflink">#</a></p><div class="subs constructors"><p class="caption">Constructors</p><table><tr><td class="src"><a id="v:FFI" class="def">FFI</a></td><td class="doc empty">&nbsp;</td></tr><tr><td colspan="2"><div class="subs fields"><p class="caption">Fields</p><ul><li><dfn class="src"><a id="v:ffi_moduleName" class="def">ffi_moduleName</a> :: ModuleName</dfn><div class="doc empty">&nbsp;</div></li><li><dfn class="src"><a id="v:ffi_values" class="def">ffi_values</a> :: [<a href="Language-PureScript-Interpreter-FFI.html#t:ForeignImport" title="Language.PureScript.Interpreter.FFI">ForeignImport</a> m]</dfn><div class="doc empty">&nbsp;</div></li></ul></div></td></tr></table></div></div><div class="top"><p class="src"><span class="keyword">data</span> <a id="t:ForeignImport" class="def">ForeignImport</a> m <a href="src/Language.PureScript.Interpreter.FFI.html#ForeignImport" class="link">Source</a> <a href="#t:ForeignImport" class="selflink">#</a></p><div class="subs constructors"><p class="caption">Constructors</p><table><tr><td class="src"><a id="v:ForeignImport" class="def">ForeignImport</a></td><td class="doc empty">&nbsp;</td></tr><tr><td colspan="2"><div class="subs fields"><p class="caption">Fields</p><ul><li><dfn class="src"><a id="v:fv_name" class="def">fv_name</a> :: Ident</dfn><div class="doc empty">&nbsp;</div></li><li><dfn class="src"><a id="v:fv_type" class="def">fv_type</a> :: SourceType</dfn><div class="doc empty">&nbsp;</div></li><li><dfn class="src"><a id="v:fv_value" class="def">fv_value</a> :: <a href="Language-PureScript-Interpreter-Types.html#t:Value" title="Language.PureScript.Interpreter.Types">Value</a> m</dfn><div class="doc empty">&nbsp;</div></li></ul></div></td></tr></table></div></div><div class="top"><p class="src"><a id="v:toEnv" class="def">toEnv</a> :: <a href="Language-PureScript-Interpreter-FFI.html#t:FFI" title="Language.PureScript.Interpreter.FFI">FFI</a> m -&gt; <a href="Language-PureScript-Interpreter-Types.html#t:Env" title="Language.PureScript.Interpreter.Types">Env</a> m <a href="src/Language.PureScript.Interpreter.FFI.html#toEnv" class="link">Source</a> <a href="#v:toEnv" class="selflink">#</a></p></div><div class="top"><p class="src"><a id="v:toExterns" class="def">toExterns</a> :: <a href="Language-PureScript-Interpreter-FFI.html#t:FFI" title="Language.PureScript.Interpreter.FFI">FFI</a> m -&gt; ExternsFile <a href="src/Language.PureScript.Interpreter.FFI.html#toExterns" class="link">Source</a> <a href="#v:toExterns" class="selflink">#</a></p></div></div></div><div id="footer"><p>Produced by <a href="http://www.haskell.org/haddock/">Haddock</a> version 2.24.0</p></div></body></html>
